update test snapshots to match schema v1.0.0

+++ tests/testthat/test-create_model_task.R
@@ -72,7 +72,7 @@ test_that("create_model_task functions work correctly", {
         ),
         create_output_type_median(
           is_required = FALSE,
-          value_type = "numeric"
+          value_type = "double"
         ),
         create_output_type_quantile(
           required = c(0.25, 0.5, 0.75),
